+++ src/Streamly/Streams/Prelude.hs
@@ -21,13 +21,8 @@
 --
 module Streamly.Streams.Prelude
     (
-    -- * Construction
-      streamBuild
-    , fromCallback
-
     -- * Elimination
-    , streamFold
-    , runStream
+      runStream
     , runStreaming      -- deprecated
 
     -- * Fold Utilities
@@ -38,10 +33,27 @@ module Streamly.Streams.Prelude
 where
 
 import Streamly.Streams.Serial (SerialT)
-import Streamly.SVar (SVar)
-
 import Streamly.Streams.StreamK hiding (runStream)
 import qualified Streamly.Streams.StreamK as K
+
+------------------------------------------------------------------------------
+-- Eliminating a stream
+------------------------------------------------------------------------------
+
+-- | Run a streaming composition, discard the results. By default it interprets
+-- the stream as 'SerialT', to run other types of streams use the type adapting
+-- combinators for example @runStream . 'asyncly'@.
+--
+-- @since 0.2.0
+runStream :: Monad m => SerialT m a -> m ()
+runStream = K.runStream
+
+-- | Same as 'runStream'
+--
+-- @since 0.1.0
+{-# DEPRECATED runStreaming "Please use runStream instead." #-}
+runStreaming :: (Monad m, IsStream t) => t m a -> m ()
+runStreaming = runStream . adapt
 
 ------------------------------------------------------------------------------
 -- Fold Utilities
@@ -79,30 +91,6 @@ forEachWith :: (IsStream t, Foldable f)
     => (t m b -> t m b -> t m b) -> f a -> (a -> t m b) -> t m b
 forEachWith f xs g = foldr (f . g) nil xs
 
-------------------------------------------------------------------------------
--- Generation
-------------------------------------------------------------------------------
-
--- | Build a stream from its church encoding.  The function passed maps
--- directly to the underlying representation of the stream type. The second
--- parameter to the function is the "yield" function yielding a value and the
--- remaining stream if any otherwise 'Nothing'. The third parameter is to
--- represent an "empty" stream.
-streamBuild :: IsStream t
-    => (forall r. Maybe (SVar Stream m a)
-        -> (a -> t m a -> m r)
-        -> (a -> m r)
-        -> m r
-        -> m r)
-    -> t m a
-streamBuild k = fromStream $ Stream $ \svr stp sng yld ->
-    let yield a r = yld a (toStream r)
-     in k svr yield sng stp
-
--- | Build a singleton stream from a callback function.
-fromCallback :: IsStream t => (forall r. (a -> m r) -> m r) -> t m a
-fromCallback k = fromStream $ Stream $ \_ _ sng _ -> k sng
-
 -------------------------------------------------------------------------------
 -- Generation by unfold
 -------------------------------------------------------------------------------
@@ -128,37 +116,3 @@ unfoldrM step = go
             Just (a, b) ->
                 K.runStream (a `cons` go b) svr stp sng yld
 -}
-
-------------------------------------------------------------------------------
--- Destroying a stream
-------------------------------------------------------------------------------
-
--- | Fold a stream using its church encoding. The second argument is the "step"
--- function consuming an element and the remaining stream, if any. The third
--- argument is for consuming an "empty" stream that yields nothing.
-streamFold
-    :: IsStream t
-    => Maybe (SVar Stream m a)
-    -> (a -> t m a -> m r)
-    -> (a -> m r)
-    -> m r
-    -> t m a
-    -> m r
-streamFold svr step single blank m =
-    let yield a x = step a (fromStream x)
-     in (K.runStream (toStream m)) svr blank single yield
-
--- | Run a streaming composition, discard the results. By default it interprets
--- the stream as 'SerialT', to run other types of streams use the type adapting
--- combinators for example @runStream . 'asyncly'@.
---
--- @since 0.2.0
-runStream :: Monad m => SerialT m a -> m ()
-runStream = run
-
--- | Same as 'runStream'
---
--- @since 0.1.0
-{-# DEPRECATED runStreaming "Please use runStream instead." #-}
-runStreaming :: (Monad m, IsStream t) => t m a -> m ()
-runStreaming = runStream . adapt