+++ src/Streamly/Streams/StreamK.hs
@@ -20,32 +20,47 @@
 -- Portability : GHC
 --
 --
+-- Continuation passing style stream implementation.
+--
+-- import qualified Streamly.Streams.StreamK as K
+--
 module Streamly.Streams.StreamK
     (
     -- * Streams
       IsStream (..)
     , adapt
-    , Streaming         -- deprecated
+
     , Stream (..)
-    , run
+    , mkStream
+    , foldStream
+    , runStream
 
     -- * Construction
     , nil
+    , yield
+    , yieldM
     , cons
     , (.:)
-    , consMSerial
+
+    -- * Asynchronous construction
+    , nilK
+    , yieldK
+    , consK
 
     -- * Generation
-    , singleton
-    , once
     , repeat
 
     -- * Semigroup Style Composition
     , serial
 
     -- * Utilities
+    , consMSerial
     , bindWith
     , withLocal
+
+    -- * Deprecated
+    , Streaming -- deprecated
+    , once      -- deprecated
     )
 where
 
@@ -77,7 +92,7 @@ import Streamly.SVar
 --
 newtype Stream m a =
     Stream {
-        runStream :: forall r.
+        unStream :: forall r.
                Maybe (SVar Stream m a)   -- local state
             -> m r                       -- stop
             -> (a -> m r)                -- singleton
@@ -153,6 +168,23 @@ adapt :: (IsStream t1, IsStream t2) => t1 m a -> t2 m a
 adapt = fromStream . toStream
 
 ------------------------------------------------------------------------------
+-- Building a stream
+------------------------------------------------------------------------------
+
+-- | Build a stream from an 'SVar', a stop continuation, a singleton stream
+-- continuation and a yield continuation.
+mkStream:: IsStream t
+    => (forall r. Maybe (SVar Stream m a)
+        -> m r
+        -> (a -> m r)
+        -> (a -> t m a -> m r)
+        -> m r)
+    -> t m a
+mkStream k = fromStream $ Stream $ \svr stp sng yld ->
+    let yieldk a r = yld a (toStream r)
+     in k svr stp sng yieldk
+
+------------------------------------------------------------------------------
 -- Construction
 ------------------------------------------------------------------------------
 
@@ -167,11 +199,41 @@ adapt = fromStream . toStream
 nil :: IsStream t => t m a
 nil = fromStream $ Stream $ \_ stp _ _ -> stp
 
+-- faster than yieldM because there is no bind.
+-- | Create a singleton stream from a pure value. Same as @yieldM (return a)@
+-- but more efficient.
+yield :: IsStream t => a -> t m a
+yield a = fromStream $ Stream $ \_ _ single _ -> single a
+
+-- | Create a singleton stream from a monadic action. Same as @m \`consM` nil@
+-- but more efficient.
+--
+-- @
+-- > toList $ yieldM getLine
+-- hello
+-- ["hello"]
+-- @
+--
+-- @since 0.4.0
+{-# INLINE yieldM #-}
+yieldM :: (Monad m, IsStream t) => m a -> t m a
+yieldM m = fromStream $ Stream $ \_ _ single _ -> m >>= single
+
+-- | Same as yieldM
+--
+-- @since 0.2.0
+{-# DEPRECATED once "Please use yieldM instead." #-}
+{-# INLINE once #-}
+once :: (Monad m, IsStream t) => m a -> t m a
+once = yieldM
+
 infixr 5 `cons`
 
 -- faster than consM because there is no bind.
 -- | Construct a stream by adding a pure value at the head of an existing
--- stream. For pure values it can be faster than 'consM'. For example:
+-- stream. For serial streams this is the same as @(return a) \`consM` r@ but
+-- more efficient. For concurrent streams this is not concurrent whereas
+-- 'consM' is concurrent. For example:
 --
 -- @
 -- > toList $ 1 \`cons` 2 \`cons` 3 \`cons` nil
@@ -210,6 +272,25 @@ infixr 5 .:
 consMSerial :: (Monad m) => m a -> Stream m a -> Stream m a
 consMSerial m r = Stream $ \_ _ _ yld -> m >>= \a -> yld a r
 
+------------------------------------------------------------------------------
+-- Asynchronous construction
+------------------------------------------------------------------------------
+
+-- | Make an empty stream from a callback function.
+nilK :: IsStream t => (forall r. m r -> m r) -> t m a
+nilK k = fromStream $ Stream $ \_ stp _ _ -> k stp
+
+-- | Make a singleton stream from a one shot callback function.
+yieldK :: IsStream t => (forall r. (a -> m r) -> m r) -> t m a
+yieldK k = fromStream $ Stream $ \_ _ sng _ -> k sng
+
+-- | Construct a stream from a callback function.
+consK :: IsStream t => (forall r. (a -> m r) -> m r) -> t m a -> t m a
+consK k r = fromStream $ Stream $ \_ _ _ yld -> k (\x -> yld x (toStream r))
+
+-- XXX consK with concurrent callbacks
+-- XXX Build a stream from a repeating callback function.
+
 -------------------------------------------------------------------------------
 -- IsStream Stream
 -------------------------------------------------------------------------------
@@ -228,36 +309,32 @@ instance IsStream Stream where
     (|:) :: Monad m => m a -> Stream m a -> Stream m a
     (|:) = consMSerial
 
-run :: (Monad m, IsStream t) => t m a -> m ()
-run m = go (toStream m)
+-- | Fold a stream by providing an SVar, a stop continuation, a singleton
+-- continuation and a yield continuation.
+foldStream
+    :: IsStream t
+    => Maybe (SVar Stream m a)
+    -> m r
+    -> (a -> m r)
+    -> (a -> t m a -> m r)
+    -> t m a
+    -> m r
+foldStream svr blank single step m =
+    let yieldk a x = step a (fromStream x)
+     in (unStream (toStream m)) svr blank single yieldk
+
+runStream :: (Monad m, IsStream t) => t m a -> m ()
+runStream m = go (toStream m)
     where
     go m1 =
         let stop = return ()
             single _ = return ()
-            yield _ r = go (toStream r)
-         in (runStream m1) Nothing stop single yield
+            yieldk _ r = go (toStream r)
+         in (unStream m1) Nothing stop single yieldk
 
 -------------------------------------------------------------------------------
 -- Special generation
 -------------------------------------------------------------------------------
-
--- | Same as @once . return@ but may be faster because there is no bind
-singleton :: IsStream t => a -> t m a
-singleton a = fromStream $ Stream $ \_ _ single _ -> single a
-
--- | Create a singleton stream by executing a monadic action once. Same as
--- @m \`consM` nil@ but more efficient.
---
--- @
--- > toList $ once getLine
--- hello
--- ["hello"]
--- @
---
--- @since 0.2.0
-{-# INLINE once #-}
-once :: (Monad m, IsStream t) => m a -> t m a
-once m = fromStream $ Stream $ \_ _ single _ -> m >>= single
 
 repeat :: IsStream t => a -> t m a
 repeat a = let x = cons a x in x
@@ -273,10 +350,10 @@ serial :: Stream m a -> Stream m a -> Stream m a
 serial m1 m2 = go m1
     where
     go (Stream m) = Stream $ \_ stp sng yld ->
-            let stop      = (runStream m2) Nothing stp sng yld
-                single a  = yld a m2
-                yield a r = yld a (go r)
-            in m Nothing stop single yield
+            let stop       = (unStream m2) Nothing stp sng yld
+                single a   = yld a m2
+                yieldk a r = yld a (go r)
+            in m Nothing stop single yieldk
 
 instance Semigroup (Stream m a) where
     (<>) = serial
@@ -295,9 +372,9 @@ instance Monoid (Stream m a) where
 
 instance Monad m => Functor (Stream m) where
     fmap f m = Stream $ \_ stp sng yld ->
-        let single    = sng . f
-            yield a r = yld (f a) (fmap f r)
-        in (runStream m) Nothing stp single yield
+        let single     = sng . f
+            yieldk a r = yld (f a) (fmap f r)
+        in (unStream m) Nothing stp single yieldk
 
 -------------------------------------------------------------------------------
 -- Bind utility
@@ -313,10 +390,10 @@ bindWith par m f = go m
     where
         go (Stream g) =
             Stream $ \ctx stp sng yld ->
-            let runIt x = (runStream x) ctx stp sng yld
-                single a  = runIt $ f a
-                yield a r = runIt $ f a `par` go r
-            in g Nothing stp single yield
+            let run x = (unStream x) ctx stp sng yld
+                single a   = run $ f a
+                yieldk a r = run $ f a `par` go r
+            in g Nothing stp single yieldk
 
 ------------------------------------------------------------------------------
 -- Alternative & MonadPlus
@@ -324,8 +401,8 @@ bindWith par m f = go m
 
 _alt :: Stream m a -> Stream m a -> Stream m a
 _alt m1 m2 = Stream $ \_ stp sng yld ->
-    let stop  = runStream m2 Nothing stp sng yld
-    in runStream m1 Nothing stop sng yld
+    let stop  = unStream m2 Nothing stp sng yld
+    in unStream m1 Nothing stop sng yld
 
 ------------------------------------------------------------------------------
 -- MonadReader
@@ -335,8 +412,8 @@ withLocal :: MonadReader r m => (r -> r) -> Stream m a -> Stream m a
 withLocal f m =
     Stream $ \_ stp sng yld ->
         let single = local f . sng
-            yield a r = local f $ yld a (withLocal f r)
-        in (runStream m) Nothing (local f stp) single yield
+            yieldk a r = local f $ yld a (withLocal f r)
+        in (unStream m) Nothing (local f stp) single yieldk
 
 ------------------------------------------------------------------------------
 -- MonadError
@@ -349,9 +426,9 @@ withCatchError
     => Stream m a -> (e -> Stream m a) -> Stream m a
 withCatchError m h =
     Stream $ \_ stp sng yld ->
-        let run x = runStream x Nothing stp sng yield
+        let run x = unStream x Nothing stp sng yieldk
             handle r = r `catchError` \e -> run $ h e
-            yield a r = yld a (withCatchError r h)
+            yieldk a r = yld a (withCatchError r h)
         in handle $ run m
 -}
 
@@ -360,4 +437,4 @@ withCatchError m h =
 -------------------------------------------------------------------------------
 
 instance MonadTrans Stream where
-    lift = once
+    lift = yieldM