+++ test/parallel-loops.hs
@@ -8,19 +8,19 @@ main = do
     hSetBuffering stdout LineBuffering
     runStream $ do
         x <- S.take 10 $ loop "A" `parallel` loop "B"
-        S.once $ myThreadId >>= putStr . show
+        S.yieldM $ myThreadId >>= putStr . show
                >> putStr " got "
                >> print x
 
     where
 
     -- we can just use
-    -- parallely $ cycle1 $ once (...)
+    -- parallely $ cycle1 $ yieldM (...)
     loop :: String -> Serial (String, Int)
     loop name = do
-        S.once $ threadDelay 1000000
-        rnd <- S.once (randomIO :: IO Int)
-        S.once $ myThreadId >>= putStr . show
+        S.yieldM $ threadDelay 1000000
+        rnd <- S.yieldM (randomIO :: IO Int)
+        S.yieldM $ myThreadId >>= putStr . show
                >> putStr " yielding "
                >> print rnd
         return (name, rnd) `parallel` loop name